+++ R/pladmm_mob_fit.R
@@ -10,9 +10,24 @@ pladmm_mob_fit <- function (y, worth, x = NULL, start = NULL, weights = NULL,
     dots <- list(...)
     if (!is.null(weights))
         warning("`weights` not yet implemented")
-    res <- pladmm_fit(y, X = worth, start = start, ...)
 
+    orderings <- as.matrix(as.rankings(y))
+    res <- pladmm_fit(orderings,
+                      X = worth, start = start, ...)
+
+    if (object){
+        # return dummy PLADMM object so works with methods e.g. vcov, AIC
+        res$x <- worth
+        res$orderings <- orderings
+        res$rank <- ncol(worth) - 1
+        class(res) <- "PLADMM"
+    }
     if (estfun) {
+        if (!object){
+            # add in extra required info as would be in PLADMM object
+            res$x <- worth
+            res$orderings <- orderings
+        }
         percomp <- estfun.PLADMM(res)
         estfun <- rowsum(as.matrix(percomp), attr(y, "index"))
     } else estfun <- NULL
@@ -46,7 +61,7 @@ estfun.PLADMM <- function(x, ...) {
     res <- matrix(nrow = n_rankings, ncol = n_coef,
                   dimnames = list(NULL, colnames(X)))
     for (j in seq_len(n_coef)[-1]){ #intercept not estimable
-        Xj <- matrix(c(X[,j], 0)[c(revorder)], nr = n_rankings)
+        Xj <- matrix(c(X[,j], 0)[c(revorder)], nrow = n_rankings)
         ## part 1: derivative of first term
         p1 <- rowSums(Xj, na.rm = TRUE)
         ## part 2: derivative of second term